src/streamlit/requirements.txt
# Requirements for Yellow Sticky Trap Streamlit App
streamlit>=1.22.0
Pillow>=9.0.0
torch>=1.12.0
numpy>=1.21.0
matplotlib>=3.5.0
ultralytics>=8.0.0
